+++ src/utils/ToolUtils.ts
@@ -46,13 +46,13 @@ export namespace ToolUtils {
     return shardCount;
   }
 
-  export async function isCopyFile(filePath: string): Promise<boolean> {
+  export async function isCopyFile(filePath: string, needUnlink?: boolean): Promise<boolean> {
     const copyExp = /[a-zA-Z0-9]+_[0-9]+.js/;
     const copyMatch = copyExp.exec(filePath);
     if (copyMatch) {
       try {
         const stat = await LibFs.stat(filePath);
-        if (stat.isFile()) {
+        if (stat.isFile() && needUnlink) {
           await LibFs.unlink(filePath);
         }
       } catch (error) {
@@ -76,10 +76,12 @@ export namespace ToolUtils {
 
   export async function copyNewFile(fileName: string,
                                     filePath: string, rootPath: string,
-                                    index: number): Promise<{ [key: string]: string }> {
+                                    index: number, needCopyFile?: boolean): Promise<{ [key: string]: string }> {
     const newFileName = `${fileName}_${index}`;
     const newFilePath = LibPath.join(rootPath, `${newFileName}.js`);
-    await fsCopy(filePath, newFilePath);
+    if (needCopyFile) {
+      await fsCopy(filePath, newFilePath);
+    }
     return { newFileName, newFilePath };
   }
